Acoommodate libplanet API changes
