+++ Lib9c.Policy/Policy/BlockPolicySource.cs
@@ -167,7 +167,6 @@ namespace Nekoyume.Blockchain.Policy
             {
                 if (blockChain
                     .GetWorldState()
-                    .GetAccountState(ReservedAddresses.LegacyAccount)
                     .GetBalance(MeadConfig.PatronAddress, Currencies.Mead) < 1 * Currencies.Mead)
                 {
                     // Check Activation
